Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/parsing/transition-timing-function-valid-expected.txt


PASS e.style['transition-timing-function'] = "linear" should set the property value 
PASS e.style['transition-timing-function'] = "ease" should set the property value 
PASS e.style['transition-timing-function'] = "ease-in" should set the property value 
PASS e.style['transition-timing-function'] = "ease-out" should set the property value 
PASS e.style['transition-timing-function'] = "ease-in-out" should set the property value 
PASS e.style['transition-timing-function'] = "cubic-bezier(0.1, 0.2, 0.8, 0.9)" should set the property value 
PASS e.style['transition-timing-function'] = "cubic-bezier(0, -2, 1, 3)" should set the property value 
PASS e.style['transition-timing-function'] = "cubic-bezier(0, 0.7, 1, 1.3)" should set the property value 
FAIL e.style['transition-timing-function'] = "step-start" should set the property value assert_equals: serialization should be canonical expected "steps(1, start)" but got "step-start"
FAIL e.style['transition-timing-function'] = "step-end" should set the property value assert_equals: serialization should be canonical expected "steps(1)" but got "step-end"
FAIL e.style['transition-timing-function'] = "steps(4)" should set the property value assert_equals: serialization should be canonical expected "steps(4)" but got "steps(4, end)"
PASS e.style['transition-timing-function'] = "steps(4, start)" should set the property value 
FAIL e.style['transition-timing-function'] = "steps(2, end)" should set the property value assert_equals: serialization should be canonical expected "steps(2)" but got "steps(2, end)"
FAIL e.style['transition-timing-function'] = "steps(2, jump-start)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['transition-timing-function'] = "steps(2, jump-end)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['transition-timing-function'] = "steps(2, jump-both)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['transition-timing-function'] = "steps(2, jump-none)" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['transition-timing-function'] = "linear, ease, linear" should set the property value 
